% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getData}
\alias{getData}
\title{API wrapper that fetches data from IsoMemo Applications}
\usage{
getData(
  db = getDatabaseList(mapping = "IsoMemo"),
  category = NULL,
  field = NULL,
  mapping = "IsoMemo"
)
}
\arguments{
\item{db}{(character) database options:  "14CSea"   "CIMA"     "IntChron" "LiVES"}

\item{category}{(character) domain specific categories of fields to retrieve: "Dating info","Isotopic proxies." If set to NULL (default) all categories are returned}

\item{field}{(character) fields to return. If set to NULL (default) all fields will be returned}

\item{mapping}{(character) Optionally, provide a specific mapping in order to obtain a list
of databases only for that mapping. Check available mapping ids with getMappings()}
}
\value{
A data frame containing the requested databases, category domains, and variables of interest from the user
}
\description{
Data retrieval for databases within the IsoMemo initiative (https://isomemo.com/). Given the parameters Database
name, category names, and field names.
}
