% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getDatabaseList}
\alias{getDatabaseList}
\title{Get Database List}
\usage{
getDatabaseList(mapping = "IsoMemo")
}
\arguments{
\item{mapping}{(character) Optionally, provide a specific mapping in order to obtain a list
of databases only for that mapping. Check available mapping ids with getMappings()}
}
\value{
A character vector with all available databases
}
\description{
Get Database List
}
