% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getDatabaseList}
\alias{getDatabaseList}
\title{Get list of databases available}
\usage{
getDatabaseList()
}
\value{
A character vector of possible database names currently available
}
\description{
Get list of databases available
}
