% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-getData.R
\name{getFields}
\alias{getFields}
\title{Get field mapping table}
\usage{
getFields()
}
\value{
A data frame that describes data field name, data type, and domain category
}
\description{
Get field mapping table
}
